requirements.txt
gspread==5.12.4
google-auth==2.27.0
openai==1.12.0
pandas==2.2.1
markdown==3.5.2
weasyprint==60.2
fpdf==1.7.2
pymupdf==1.23.26
python-dotenv==1.0.1 